requirements.txt
mysql-connector-python
pandas
matplotlib
seaborn
scikit-learn
imbalanced-learn
streamlit
sqlalchemy
joblib